refactor: rename components to index.tsx to simplify imports
