+++ src/components/button/Button.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { describe, it, expect, vi } from "vitest";
-import Button from "./Button";
+import Button from "./";
 
 describe("Button", () => {
   it("renders button with provided text", () => {
